Split incoming processing from store.rs to a new incoming.rs
