+++ components/places/src/bookmark_sync/mod.rs
@@ -2,6 +2,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
+mod incoming;
 pub mod record;
 pub mod store;
